DSL::validate - force explicit file/string designation
Outthentic::DSL->validate took 1 argument, if a file existed with that
name, it'd load that file line-wise and use the lines as a list of
checks.  Otherwise it'd either treat it as an arrayref of checks or a
single check.

This has a couple issues but the major one is that the argument is
ambiguous - if a program were calling $test->validate("OK"), an empty
file could be created in the working directory called 'OK' and cause
entirely different behavior.

Removing this ambiguity is not a change that can be done in a
backwards-compatible way.

If ->validate is passed a single argument, it will now warn with
instructions to use the proper two-argument syntax:

files:
  ...->validate(from_file => $filename);

string:
  ...->validate(from_string => $str);